Restructure code and add jsdoc #13 #4
