+++ index.html
@@ -16,10 +16,12 @@
     <br>
     <br>
     <br>
-    <img id="ball" src="image/8ball_image.png">
+    <img id="ball" src="./assets/image/8ball_image.png">
     <p id="answer"></p>
   </div>
-  <script src="script.js"></script>
+
+  <script type="module" src="./assets/js/main.js"></script>
+  <script type="module" src="./assets/js/dom.js"></script>
 </body>
 
 </html>
